+++ scripts/macros.js
@@ -191,8 +191,7 @@ export {channelDivinity} from './macros/2024/classFeatures/cleric/channelDivinit
 export {searUndead} from './macros/2024/classFeatures/cleric/searUndead.js';
 export {divineStrike} from './macros/2024/classFeatures/cleric/divineStrike.js';
 export {potentSpellcasting} from './macros/2024/classFeatures/cleric/potentSpellcasting.js';
-export {improvedDivineStrike} from './macros/2024/classFeatures/cleric/improvedDivineStrike.js';
-export {improvedPotentSpellcasting} from './macros/2024/classFeatures/cleric/improvedPotentSpellcasting.js';
+export {improvedBlessedStrikes} from './macros/2024/classFeatures/cleric/improvedBlessedStrikes.js';
 export {divineIntervention} from './macros/2024/classFeatures/cleric/divineIntervention.js';
 export {greaterDivineIntervention, greaterDivineInterventionRest} from './macros/2024/classFeatures/cleric/greaterDivineIntervention.js';
 //Life Cleric