+++ scripts/macros/2024/classFeatures/cleric/improvedDivineStrike.js
@@ -1,5 +1,0 @@
-export let improvedDivineStrike = {
-    name: 'Improved Blessed Strikes: Divine Strike',
-    version: '1.2.12',
-    rules: 'modern'
-};